refactor: add the slow query timer in the `Tikcet` of ProcessManager
