+++ src/catalog/Cargo.toml
@@ -45,6 +45,7 @@ partition.workspace = true
 paste.workspace = true
 prometheus.workspace = true
 promql-parser.workspace = true
+rand.workspace = true
 rustc-hash.workspace = true
 serde_json.workspace = true
 session.workspace = true
