+++ src/catalog/src/process_manager.rs
@@ -17,16 +17,21 @@ use std::collections::HashMap;
 use std::fmt::{Debug, Display, Formatter};
 use std::sync::atomic::{AtomicU32, Ordering};
 use std::sync::{Arc, RwLock};
+use std::time::{Duration, Instant, UNIX_EPOCH};
 
 use api::v1::frontend::{KillProcessRequest, ListProcessRequest, ProcessInfo};
 use common_base::cancellation::CancellationHandle;
 use common_frontend::selector::{FrontendSelector, MetaClientSelector};
-use common_telemetry::{debug, info, warn};
+use common_frontend::slow_query_event::SlowQueryEvent;
+use common_telemetry::{debug, error, info, warn};
 use common_time::util::current_time_millis;
 use meta_client::MetaClientRef;
 use promql_parser::parser::EvalStmt;
+use rand::random;
+use session::context::QueryContextRef;
 use snafu::{ensure, OptionExt, ResultExt};
 use sql::statements::statement::Statement;
+use tokio::sync::mpsc::Sender;
 
 use crate::error;
 use crate::metrics::{PROCESS_KILL_COUNT, PROCESS_LIST_COUNT};
@@ -83,16 +88,17 @@ impl ProcessManager {
         self: &Arc<Self>,
         catalog: String,
         schemas: Vec<String>,
-        query: QueryStatement,
+        query: String,
         client: String,
         query_id: Option<ProcessId>,
+        _slow_query_timer: Option<SlowQueryTimer>,
     ) -> Ticket {
         let id = query_id.unwrap_or_else(|| self.next_id.fetch_add(1, Ordering::Relaxed));
         let process = ProcessInfo {
             id,
             catalog: catalog.clone(),
             schemas,
-            query: query.to_string(),
+            query,
             start_timestamp: current_time_millis(),
             client,
             frontend: self.server_addr.clone(),
@@ -112,6 +118,7 @@ impl ProcessManager {
             manager: self.clone(),
             id,
             cancellation_handle,
+            _slow_query_timer,
         }
     }
 
@@ -242,6 +249,7 @@ pub struct Ticket {
     pub(crate) manager: ProcessManagerRef,
     pub(crate) id: ProcessId,
     pub cancellation_handle: Arc<CancellationHandle>,
+    _slow_query_timer: Option<SlowQueryTimer>,
 }
 
 impl Drop for Ticket {
@@ -282,35 +290,122 @@ impl Debug for CancellableProcess {
     }
 }
 
+/// SlowQueryTimer is used to log slow query when it's dropped.
+/// In drop(), it will check if the query is slow and send the slow query event to the handler.
+pub struct SlowQueryTimer {
+    start: Instant,
+    stmt: QueryStatement,
+    query_ctx: QueryContextRef,
+    threshold: Option<Duration>,
+    sample_ratio: Option<f64>,
+    tx: Sender<SlowQueryEvent>,
+}
+
+impl SlowQueryTimer {
+    pub fn new(
+        stmt: QueryStatement,
+        query_ctx: QueryContextRef,
+        threshold: Option<Duration>,
+        sample_ratio: Option<f64>,
+        tx: Sender<SlowQueryEvent>,
+    ) -> Self {
+        Self {
+            start: Instant::now(),
+            stmt,
+            query_ctx,
+            threshold,
+            sample_ratio,
+            tx,
+        }
+    }
+}
+
+impl SlowQueryTimer {
+    fn send_slow_query_event(&self, elapsed: Duration, threshold: Duration) {
+        let mut slow_query_event = SlowQueryEvent {
+            cost: elapsed.as_millis() as u64,
+            threshold: threshold.as_millis() as u64,
+            query: "".to_string(),
+            query_ctx: self.query_ctx.clone(),
+
+            // The following fields are only used for PromQL queries.
+            is_promql: false,
+            promql_range: None,
+            promql_step: None,
+            promql_start: None,
+            promql_end: None,
+        };
+
+        match &self.stmt {
+            QueryStatement::Promql(stmt) => {
+                slow_query_event.is_promql = true;
+                slow_query_event.query = stmt.expr.to_string();
+                slow_query_event.promql_step = Some(stmt.interval.as_millis() as u64);
+
+                let start = stmt
+                    .start
+                    .duration_since(UNIX_EPOCH)
+                    .unwrap_or_default()
+                    .as_millis() as i64;
+
+                let end = stmt
+                    .end
+                    .duration_since(UNIX_EPOCH)
+                    .unwrap_or_default()
+                    .as_millis() as i64;
+
+                slow_query_event.promql_range = Some((end - start) as u64);
+                slow_query_event.promql_start = Some(start);
+                slow_query_event.promql_end = Some(end);
+            }
+            QueryStatement::Sql(stmt) => {
+                slow_query_event.query = stmt.to_string();
+            }
+        }
+
+        // Send SlowQueryEvent to the handler.
+        if let Err(e) = self.tx.try_send(slow_query_event) {
+            error!(e; "Failed to send slow query event");
+        }
+    }
+}
+
+impl Drop for SlowQueryTimer {
+    fn drop(&mut self) {
+        if let Some(threshold) = self.threshold {
+            // Calculate the elaspsed duration since the timer is created.
+            let elapsed = self.start.elapsed();
+            if elapsed > threshold {
+                if let Some(ratio) = self.sample_ratio {
+                    // Only capture a portion of slow queries based on sample_ratio.
+                    // Generate a random number in [0, 1) and compare it with sample_ratio.
+                    if ratio >= 1.0 || random::<f64>() <= ratio {
+                        self.send_slow_query_event(elapsed, threshold);
+                    }
+                } else {
+                    // Captures all slow queries if sample_ratio is not set.
+                    self.send_slow_query_event(elapsed, threshold);
+                }
+            }
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use std::sync::Arc;
 
-    use sql::dialect::GreptimeDbDialect;
-    use sql::parser::{ParseOptions, ParserContext};
-
-    use super::QueryStatement;
     use crate::process_manager::ProcessManager;
 
     #[tokio::test]
     async fn test_register_query() {
         let process_manager = Arc::new(ProcessManager::new("127.0.0.1:8000".to_string(), None));
-        let sql_query = "SELECT * FROM test_table";
-        let sql_query_statement = QueryStatement::Sql(
-            ParserContext::create_with_dialect(
-                sql_query,
-                &GreptimeDbDialect {},
-                ParseOptions::default(),
-            )
-            .unwrap()[0]
-                .clone(),
-        );
-
         let ticket = process_manager.clone().register_query(
             "public".to_string(),
             vec!["test".to_string()],
-            sql_query_statement,
+            "SELECT * FROM table".to_string(),
             "".to_string(),
+            None,
             None,
         );
 
@@ -318,7 +413,7 @@ mod tests {
         assert_eq!(running_processes.len(), 1);
         assert_eq!(&running_processes[0].frontend, "127.0.0.1:8000");
         assert_eq!(running_processes[0].id, ticket.id);
-        assert_eq!(&running_processes[0].query, sql_query);
+        assert_eq!(&running_processes[0].query, "SELECT * FROM table");
 
         drop(ticket);
         assert_eq!(process_manager.local_processes(None).unwrap().len(), 0);
@@ -328,22 +423,14 @@ mod tests {
     async fn test_register_query_with_custom_id() {
         let process_manager = Arc::new(ProcessManager::new("127.0.0.1:8000".to_string(), None));
         let custom_id = 12345;
-        let sql_query_statement = QueryStatement::Sql(
-            ParserContext::create_with_dialect(
-                "SELECT * FROM test_table",
-                &GreptimeDbDialect {},
-                ParseOptions::default(),
-            )
-            .unwrap()[0]
-                .clone(),
-        );
 
         let ticket = process_manager.clone().register_query(
             "public".to_string(),
             vec!["test".to_string()],
-            sql_query_statement,
+            "SELECT * FROM table".to_string(),
             "client1".to_string(),
             Some(custom_id),
+            None,
         );
 
         assert_eq!(ticket.id, custom_id);
@@ -361,32 +448,18 @@ mod tests {
         let ticket1 = process_manager.clone().register_query(
             "public".to_string(),
             vec!["schema1".to_string()],
-            QueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM table1",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "SELECT * FROM table1".to_string(),
             "client1".to_string(),
+            None,
             None,
         );
 
         let ticket2 = process_manager.clone().register_query(
             "public".to_string(),
             vec!["schema2".to_string()],
-            QueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM table2",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "SELECT * FROM table2".to_string(),
             "client2".to_string(),
+            None,
             None,
         );
 
@@ -406,32 +479,18 @@ mod tests {
         let _ticket1 = process_manager.clone().register_query(
             "catalog1".to_string(),
             vec!["schema1".to_string()],
-            QueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM table1",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "SELECT * FROM table1".to_string(),
             "client1".to_string(),
+            None,
             None,
         );
 
         let _ticket2 = process_manager.clone().register_query(
             "catalog2".to_string(),
             vec!["schema2".to_string()],
-            QueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM table2",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "SELECT * FROM table2".to_string(),
             "client2".to_string(),
+            None,
             None,
         );
 
@@ -456,16 +515,9 @@ mod tests {
         let ticket = process_manager.clone().register_query(
             "public".to_string(),
             vec!["test".to_string()],
-            QueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM test_table",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "SELECT * FROM table".to_string(),
             "client1".to_string(),
+            None,
             None,
         );
         assert_eq!(process_manager.local_processes(None).unwrap().len(), 1);
@@ -480,16 +532,9 @@ mod tests {
         let ticket = process_manager.clone().register_query(
             "public".to_string(),
             vec!["test".to_string()],
-            QueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM test_table",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "SELECT * FROM table".to_string(),
             "client1".to_string(),
+            None,
             None,
         );
 
@@ -505,16 +550,9 @@ mod tests {
         let ticket = process_manager.clone().register_query(
             "public".to_string(),
             vec!["test".to_string()],
-            QueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM test_table",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "SELECT * FROM table".to_string(),
             "client1".to_string(),
+            None,
             None,
         );
         assert!(!ticket.cancellation_handle.is_cancelled());
@@ -558,17 +596,10 @@ mod tests {
         let _ticket = process_manager.clone().register_query(
             "test_catalog".to_string(),
             vec!["schema1".to_string(), "schema2".to_string()],
-            QueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT COUNT(*) FROM users WHERE age > 18",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "SELECT COUNT(*) FROM users WHERE age > 18".to_string(),
             "test_client".to_string(),
             Some(42),
+            None,
         );
 
         let processes = process_manager.local_processes(None).unwrap();
@@ -592,16 +623,9 @@ mod tests {
             let _ticket = process_manager.clone().register_query(
                 "public".to_string(),
                 vec!["test".to_string()],
-                QueryStatement::Sql(
-                    ParserContext::create_with_dialect(
-                        "SELECT * FROM test_table",
-                        &GreptimeDbDialect {},
-                        ParseOptions::default(),
-                    )
-                    .unwrap()[0]
-                        .clone(),
-                ),
+                "SELECT * FROM table".to_string(),
                 "client1".to_string(),
+                None,
                 None,
             );
