+++ src/frontend/src/instance.rs
@@ -220,8 +220,8 @@ impl Instance {
         let query_interceptor = self.plugins.get::<SqlQueryInterceptorRef<Error>>();
         let query_interceptor = query_interceptor.as_ref();
 
-        let _slow_query_timer = if let Some(recorder) = &self.slow_query_recorder {
-            recorder.start(QueryStatement::Sql(stmt.clone()), query_ctx.clone())
+        let slow_query_timer = if let Some(recorder) = &self.slow_query_recorder {
+            recorder.start(CatalogQueryStatement::Sql(stmt.clone()), query_ctx.clone())
         } else {
             None
         };
@@ -229,9 +229,10 @@ impl Instance {
         let ticket = self.process_manager.register_query(
             query_ctx.current_catalog().to_string(),
             vec![query_ctx.current_schema()],
-            CatalogQueryStatement::Sql(stmt.clone()),
+            stmt.to_string(),
             query_ctx.conn_info().to_string(),
             Some(query_ctx.process_id()),
+            slow_query_timer,
         );
 
         let query_fut = self.exec_statement_with_timeout(stmt, query_ctx, query_interceptor);
@@ -671,12 +672,6 @@ impl PrometheusHandler for Instance {
             }
         })?;
 
-        let _slow_query_timer = if let Some(recorder) = &self.slow_query_recorder {
-            recorder.start(stmt.clone(), query_ctx.clone())
-        } else {
-            None
-        };
-
         let plan = self
             .statement_executor
             .plan(&stmt, query_ctx.clone())
@@ -697,12 +692,19 @@ impl PrometheusHandler for Instance {
             .fail();
         };
 
+        let slow_query_timer = if let Some(recorder) = &self.slow_query_recorder {
+            recorder.start(query_statment.clone(), query_ctx.clone())
+        } else {
+            None
+        };
+
         let ticket = self.process_manager.register_query(
             query_ctx.current_catalog().to_string(),
             vec![query_ctx.current_schema()],
-            query_statment,
+            query_statment.to_string(),
             query_ctx.conn_info().to_string(),
             Some(query_ctx.process_id()),
+            slow_query_timer,
         );
 
         let query_fut = self.statement_executor.exec_plan(plan, query_ctx.clone());
