+++ src/frontend/src/stream_wrapper.rs
@@ -81,7 +81,7 @@ mod tests {
     use std::task::{Context, Poll};
     use std::time::Duration;
 
-    use catalog::process_manager::{ProcessManager, QueryStatement as CatalogQueryStatement};
+    use catalog::process_manager::ProcessManager;
     use common_recordbatch::adapter::RecordBatchMetrics;
     use common_recordbatch::{OrderOption, RecordBatch, RecordBatchStream};
     use datatypes::data_type::ConcreteDataType;
@@ -89,8 +89,6 @@ mod tests {
     use datatypes::schema::{ColumnSchema, Schema, SchemaRef};
     use datatypes::vectors::Int32Vector;
     use futures::{Stream, StreamExt};
-    use sql::dialect::GreptimeDbDialect;
-    use sql::parser::{ParseOptions, ParserContext};
     use tokio::time::{sleep, timeout};
 
     use super::CancellableStreamWrapper;
@@ -186,16 +184,9 @@ mod tests {
         let ticket = process_manager.register_query(
             "catalog".to_string(),
             vec![],
-            CatalogQueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM test_table",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "query".to_string(),
             "client".to_string(),
+            None,
             None,
         );
 
@@ -217,16 +208,9 @@ mod tests {
         let ticket = process_manager.register_query(
             "catalog".to_string(),
             vec![],
-            CatalogQueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM test_table",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "query".to_string(),
             "client".to_string(),
+            None,
             None,
         );
 
@@ -249,16 +233,9 @@ mod tests {
         let ticket = process_manager.register_query(
             "catalog".to_string(),
             vec![],
-            CatalogQueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM test_table",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "query".to_string(),
             "client".to_string(),
+            None,
             None,
         );
         let cancellation_handle = ticket.cancellation_handle.clone();
@@ -284,16 +261,9 @@ mod tests {
         let ticket = process_manager.register_query(
             "catalog".to_string(),
             vec![],
-            CatalogQueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM test_table",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "query".to_string(),
             "client".to_string(),
+            None,
             None,
         );
         let cancellation_handle = ticket.cancellation_handle.clone();
@@ -320,16 +290,9 @@ mod tests {
         let ticket = process_manager.register_query(
             "catalog".to_string(),
             vec![],
-            CatalogQueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM test_table",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "query".to_string(),
             "client".to_string(),
+            None,
             None,
         );
 
@@ -358,16 +321,9 @@ mod tests {
         let ticket = process_manager.register_query(
             "catalog".to_string(),
             vec![],
-            CatalogQueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM test_table",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "query".to_string(),
             "client".to_string(),
+            None,
             None,
         );
 
@@ -394,16 +350,9 @@ mod tests {
         let ticket = process_manager.register_query(
             "catalog".to_string(),
             vec![],
-            CatalogQueryStatement::Sql(
-                ParserContext::create_with_dialect(
-                    "SELECT * FROM test_table",
-                    &GreptimeDbDialect {},
-                    ParseOptions::default(),
-                )
-                .unwrap()[0]
-                    .clone(),
-            ),
+            "query".to_string(),
             "client".to_string(),
+            None,
             None,
         );
         let cancellation_handle = ticket.cancellation_handle.clone();
